task: Conditionally hide/show members directory on People page
