--- americanhandelsociety_app/templates/members_directory.html
@@ -1,12 +1,0 @@
-{% extends "base.html" %}
-
-{% block content %}
-<h1>Members</h1>
-<ul>
-{% for member in object_list %}
-    <li>{{ member.email }}</li>
-{% empty %}
-    <li>No members yet.</li>
-{% endfor %}
-</ul>
-{% endblock %}
