+++ americanhandelsociety_app/templates/partials/nav_bar.html
@@ -34,7 +34,6 @@
 					<a class="nav-link" href="{% url 'login' %}">Donate</a>
 				</li>
 
-
 				{% if request.user.is_authenticated %}
 					<li class="nav-item nav-link">|</li>
 					<li class="nav-item">
