+++ americanhandelsociety_app/templates/people.html
@@ -8,13 +8,13 @@
 <div class='container-fluid'>
 	<!-- banner -->
 	<div class="row">
-			<div class='col-sm-8 page-banner-quote'>
-				<p class="apple-pie">"Oh! Let me know when you will do that, and I will advertise it. For I am sure more people will come to see you jump, than to hear you sing."
-				</p>
-				<p>
-						George Frideric Handel
-				</p>
-			</div>
+		<div class='col-sm-8 page-banner-quote'>
+			<p class="apple-pie">"Oh! Let me know when you will do that, and I will advertise it. For I am sure more people will come to see you jump, than to hear you sing."
+			</p>
+			<p>
+					George Frideric Handel
+			</p>
+		</div>
 		<div class='col-sm-4 page-banner-header'>
 			<h1>People</h1>
 		</div>
@@ -43,5 +43,21 @@
 			<p class="text-muted"><small><span class="red-font">*</span> Honorary Director</small></p>
 		</div>
 	</div>
+
+	{% if request.user.is_authenticated %}
+		<div class="row">
+			<div class='col-12'>
+				<h2>Members Directory</h2>
+				<ul>
+				{% for member in object_list %}
+					<li>{{ member.email }}</li>
+				{% empty %}
+					<li>No members yet.</li>
+				{% endfor %}
+				</ul>
+			</div>
+		</div>
+	{% endif %}
+
 </div>
 {% endblock %}
